Access check

+++ src/services/auth.py
@@ -1,3 +1,4 @@
+import time
 from functools import lru_cache
 
 from async_fastapi_jwt_auth import AuthJWT
@@ -15,6 +16,7 @@ from db.redis import get_redis
 from models.schemas import User, Role, LoginHistory, Token
 from models.users import UserCreate, UserLogin, UserSuccessLogin
 from services.abstract import AbstractService, DeleteAbstractService
+from services.common.access_check_common import AccessCheckCommon
 
 
 class SignUpService(AbstractService):
@@ -73,7 +75,7 @@ class LoginService(AbstractService):
     async def get_data(self, user: UserLogin, user_agent: str):
         user_found = await self.get_by_login(user.login)
         if not user_found:
-            return None
+            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="User not found")
         await self.check_password(user)
         refresh_token = await self._authorize.create_refresh_token(subject=str(user_found.id))
         access_token = await self._authorize.create_access_token(subject=str(user_found.id),
@@ -115,7 +117,7 @@ class LoginService(AbstractService):
         await self._db.commit()
 
 
-class LogoutService(DeleteAbstractService):
+class LogoutService(DeleteAbstractService, AccessCheckCommon):
     def __init__(self, authorize: AuthJWT,
                  redis_token: Redis,
                  db: AsyncSession):
@@ -127,12 +129,26 @@ class LogoutService(DeleteAbstractService):
         await self._authorize.jwt_required()
 
         jwt_subject = await self._authorize.get_jwt_subject()
+
         access_token = request.cookies.get('access_token_cookie')
         refresh_token = request.cookies.get('refresh_token_cookie')
 
+        refresh_jti = await self._authorize.get_jti(refresh_token)
+        access_jti = await self._authorize.get_jti(access_token)
+
+        if await self._redis_token.exists(access_jti) or await self._redis_token.exists(refresh_jti):
+            raise HTTPException(status_code=403, detail="Access and refresh tokens are invalid")
+
         await self._authorize.unset_jwt_cookies()
 
-        await self._redis_token.set(name=access_token, value=jwt_subject, ex=900)
+        access_token_exp = (await self._authorize.get_raw_jwt(access_token))["exp"]
+        refresh_token_exp = (await self._authorize.get_raw_jwt(refresh_token))["exp"]
+            
+        access_token_ttl = access_token_exp - int(time.time())
+        refresh_token_ttl = refresh_token_exp - int(time.time())
+
+        await self._redis_token.set(name=access_jti, value=jwt_subject, ex=access_token_ttl)
+        await self._redis_token.set(name=refresh_jti, value=jwt_subject, ex=refresh_token_ttl)
         await self.delete_refresh_token_from_db(refresh_token)
 
     async def delete_refresh_token_from_db(self, refresh_token: str):
